A first try to configure the repo for building a debian package.
The complex part is that on one side a linux module should be
created with dkms. On the other sider there is a binary which
needs to be compiled on package install because it depends on
the installed hardware. For compilation the user has to provide
some additional information, which is fetched via config-package
questions and answers.
Besides this some small test and calibration binaries are created
during the package build process.

Could only test this with my single hardware (KeDei) and hope
that it works similar with the other types of displays/touchscreens.